refactor(package): rename com.nnk.springboot → com.poseidon.tradingapp
