+++ README.md
@@ -32,10 +32,10 @@
 5. Run sql script to create table doc/data.sql
 
 ## Implement a Feature
-1. Create mapping domain class and place in package com.nnk.springboot.domain
-2. Create repository class and place in package com.nnk.springboot.repositories
-3. Create controller class and place in package com.nnk.springboot.controllers
+1. Create mapping domain class and place in package com.poseidon.tradingapp.domain
+2. Create repository class and place in package com.poseidon.tradingapp.repositories
+3. Create controller class and place in package com.poseidon.tradingapp.controllers
 
 ## Security
-1. Create user service to load user from  database and place in package com.nnk.springboot.services
-2. Add configuration class and place in package com.nnk.springboot.config
+1. Create user service to load user from  database and place in package com.poseidon.tradingapp.services
+2. Add configuration class and place in package com.poseidon.tradingapp.config
