--- src/main/java/com/nnk/springboot/repositories/CurvePointRepository.java
@@ -1,9 +1,0 @@
-package com.nnk.springboot.repositories;
-
-import com.nnk.springboot.domain.CurvePoint;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface CurvePointRepository extends JpaRepository<CurvePoint, Integer> {
-
-}
