+++ src/main/java/com/poseidon/tradingapp/controllers/RuleNameController.java
@@ -1,0 +1,54 @@
+package com.poseidon.tradingapp.controllers;
+
+import com.poseidon.tradingapp.domain.RuleName;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+import jakarta.validation.Valid;
+
+@Controller
+public class RuleNameController {
+    // TODO: Inject RuleName service
+
+    @RequestMapping("/ruleName/list")
+    public String home(Model model)
+    {
+        // TODO: find all RuleName, add to model
+        return "ruleName/list";
+    }
+
+    @GetMapping("/ruleName/add")
+    public String addRuleForm(RuleName bid) {
+        return "ruleName/add";
+    }
+
+    @PostMapping("/ruleName/validate")
+    public String validate(@Valid RuleName ruleName, BindingResult result, Model model) {
+        // TODO: check data valid and save to db, after saving return RuleName list
+        return "ruleName/add";
+    }
+
+    @GetMapping("/ruleName/update/{id}")
+    public String showUpdateForm(@PathVariable("id") Integer id, Model model) {
+        // TODO: get RuleName by Id and to model then show to the form
+        return "ruleName/update";
+    }
+
+    @PostMapping("/ruleName/update/{id}")
+    public String updateRuleName(@PathVariable("id") Integer id, @Valid RuleName ruleName,
+                             BindingResult result, Model model) {
+        // TODO: check required fields, if valid call service to update RuleName and return RuleName list
+        return "redirect:/ruleName/list";
+    }
+
+    @GetMapping("/ruleName/delete/{id}")
+    public String deleteRuleName(@PathVariable("id") Integer id, Model model) {
+        // TODO: Find RuleName by Id and delete the RuleName, return to Rule list
+        return "redirect:/ruleName/list";
+    }
+}
