+++ src/main/java/com/poseidon/tradingapp/controllers/TradeController.java
@@ -1,0 +1,54 @@
+package com.poseidon.tradingapp.controllers;
+
+import com.poseidon.tradingapp.domain.Trade;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+import jakarta.validation.Valid;
+
+@Controller
+public class TradeController {
+    // TODO: Inject Trade service
+
+    @RequestMapping("/trade/list")
+    public String home(Model model)
+    {
+        // TODO: find all Trade, add to model
+        return "trade/list";
+    }
+
+    @GetMapping("/trade/add")
+    public String addUser(Trade bid) {
+        return "trade/add";
+    }
+
+    @PostMapping("/trade/validate")
+    public String validate(@Valid Trade trade, BindingResult result, Model model) {
+        // TODO: check data valid and save to db, after saving return Trade list
+        return "trade/add";
+    }
+
+    @GetMapping("/trade/update/{id}")
+    public String showUpdateForm(@PathVariable("id") Integer id, Model model) {
+        // TODO: get Trade by Id and to model then show to the form
+        return "trade/update";
+    }
+
+    @PostMapping("/trade/update/{id}")
+    public String updateTrade(@PathVariable("id") Integer id, @Valid Trade trade,
+                             BindingResult result, Model model) {
+        // TODO: check required fields, if valid call service to update Trade and return Trade list
+        return "redirect:/trade/list";
+    }
+
+    @GetMapping("/trade/delete/{id}")
+    public String deleteTrade(@PathVariable("id") Integer id, Model model) {
+        // TODO: Find Trade by Id and delete the Trade, return to Trade list
+        return "redirect:/trade/list";
+    }
+}
