+++ src/main/java/com/poseidon/tradingapp/domain/Rating.java
@@ -1,0 +1,34 @@
+package com.poseidon.tradingapp.domain;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+/**
+ * Entité représentant la notation (Rating) attribuée à un instrument financier.
+ * Chaque enregistrement contient les différentes notations selon les agences Moody's, S&P et Fitch.
+ */
+@Getter
+@Setter
+@NoArgsConstructor
+@Entity
+@Table(name = "rating")
+public class Rating {
+    @Id
+    @GeneratedValue (strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    private String moodysRating;
+    private String sandPRating;
+    private String fitchRating;
+    private Integer orderNumber;
+
+    // Constructeur pratique pour les tests et initialisations rapides
+    public Rating(String moodysRating, String sandPRating, String fitchRating, Integer orderNumber) {
+        this.moodysRating = moodysRating;
+        this.sandPRating = sandPRating;
+        this.fitchRating = fitchRating;
+        this.orderNumber = orderNumber;
+    }
+}
