+++ src/main/java/com/poseidon/tradingapp/domain/RuleName.java
@@ -1,0 +1,38 @@
+package com.poseidon.tradingapp.domain;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+/**
+ * Entité représentant une règle de validation (RuleName)
+ * utilisée dans l'application Poseidon Capital Solutions.
+ * Chaque règle contient un nom, une description et des expressions 'SQL/JSON' associées.
+ */
+@Getter
+@Setter
+@NoArgsConstructor
+@Entity
+@Table(name = "rulename")
+public class RuleName {
+    @Id
+    @GeneratedValue (strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    private String name;
+    private String description;
+    private String json;
+    private String template;
+    private String sqlStr;
+    private String sqlPart;
+
+    public RuleName(String name, String description, String json, String template, String sqlStr, String sqlPart) {
+        this.name = name;
+        this.description = description;
+        this.json = json;
+        this.template = template;
+        this.sqlStr = sqlStr;
+        this.sqlPart = sqlPart;
+    }
+}
