+++ src/main/java/com/poseidon/tradingapp/repositories/CurvePointRepository.java
@@ -1,0 +1,9 @@
+package com.poseidon.tradingapp.repositories;
+
+import com.poseidon.tradingapp.domain.CurvePoint;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface CurvePointRepository extends JpaRepository<CurvePoint, Integer> {
+
+}
