+++ src/main/java/com/poseidon/tradingapp/repositories/RuleNameRepository.java
@@ -1,0 +1,8 @@
+package com.poseidon.tradingapp.repositories;
+
+import com.poseidon.tradingapp.domain.RuleName;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface RuleNameRepository extends JpaRepository<RuleName, Integer> {
+}
