+++ src/main/java/com/poseidon/tradingapp/repositories/UserRepository.java
@@ -1,0 +1,10 @@
+package com.poseidon.tradingapp.repositories;
+
+import com.poseidon.tradingapp.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+
+
+public interface UserRepository extends JpaRepository<User, Integer>, JpaSpecificationExecutor<User> {
+
+}
