--- src/test/java/com/nnk/springboot/BidTests.java
@@ -1,47 +1,0 @@
-package com.nnk.springboot;
-
-import com.nnk.springboot.domain.BidList;
-import com.nnk.springboot.repositories.BidListRepository;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-import java.util.List;
-import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-@ExtendWith(SpringExtension.class)
-@SpringBootTest
-public class BidTests {
-
-	@Autowired
-	private BidListRepository bidListRepository;
-
-	@Test
-	public void bidListTest() {
-		BidList bid = new BidList("Account Test", "Type Test", 10d);
-
-		// Save
-		bid = bidListRepository.save(bid);
-		assertNotNull(bid.getBidListId());
-		assertEquals(10d, bid.getBidQuantity(), 10d);
-
-		// Update
-		bid.setBidQuantity(20d);
-		bid = bidListRepository.save(bid);
-		assertEquals(20d, bid.getBidQuantity(), 20d);
-
-		// Find
-		List<BidList> listResult = bidListRepository.findAll();
-        assertFalse(listResult.isEmpty());
-
-		// Delete
-		Integer id = bid.getBidListId();
-		bidListRepository.delete(bid);
-		Optional<BidList> bidList = bidListRepository.findById(id);
-		assertFalse(bidList.isPresent());
-	}
-}
