--- src/test/java/com/nnk/springboot/CurvePointTests.java
@@ -1,48 +1,0 @@
-package com.nnk.springboot;
-
-import com.nnk.springboot.domain.CurvePoint;
-import com.nnk.springboot.repositories.CurvePointRepository;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-import java.util.List;
-import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-@ExtendWith(SpringExtension.class)
-@SpringBootTest
-public class CurvePointTests {
-
-	@Autowired
-	private CurvePointRepository curvePointRepository;
-
-	@Test
-	public void curvePointTest() {
-		CurvePoint curvePoint = new CurvePoint(10, 10d, 30d);
-
-		// Save
-		curvePoint = curvePointRepository.save(curvePoint);
-		assertNotNull(curvePoint.getId());
-        assertEquals(10, (int) curvePoint.getCurveId());
-
-		// Update
-		curvePoint.setCurveId(20);
-		curvePoint = curvePointRepository.save(curvePoint);
-        assertEquals(20, (int) curvePoint.getCurveId());
-
-		// Find
-		List<CurvePoint> listResult = curvePointRepository.findAll();
-        assertFalse(listResult.isEmpty());
-
-		// Delete
-		Integer id = curvePoint.getId();
-		curvePointRepository.delete(curvePoint);
-		Optional<CurvePoint> curvePointList = curvePointRepository.findById(id);
-		assertFalse(curvePointList.isPresent());
-	}
-
-}
