--- src/test/java/com/nnk/springboot/RatingTests.java
@@ -1,47 +1,0 @@
-package com.nnk.springboot;
-
-import com.nnk.springboot.domain.Rating;
-import com.nnk.springboot.repositories.RatingRepository;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-import java.util.List;
-import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-@ExtendWith(SpringExtension.class)
-@SpringBootTest
-public class RatingTests {
-
-	@Autowired
-	private RatingRepository ratingRepository;
-
-	@Test
-	public void ratingTest() {
-		Rating rating = new Rating("Moodys Rating", "Sand PRating", "Fitch Rating", 10);
-
-		// Save
-		rating = ratingRepository.save(rating);
-		assertNotNull(rating.getId());
-        assertEquals(10, (int) rating.getOrderNumber());
-
-		// Update
-		rating.setOrderNumber(20);
-		rating = ratingRepository.save(rating);
-        assertEquals(20, (int) rating.getOrderNumber());
-
-		// Find
-		List<Rating> listResult = ratingRepository.findAll();
-        assertFalse(listResult.isEmpty());
-
-		// Delete
-		Integer id = rating.getId();
-		ratingRepository.delete(rating);
-		Optional<Rating> ratingList = ratingRepository.findById(id);
-		assertFalse(ratingList.isPresent());
-	}
-}
