+++ src/test/java/com/poseidon/tradingapp/PasswordEncodeTest.java
@@ -1,0 +1,24 @@
+package com.poseidon.tradingapp;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+/**
+ * Created by Khang Nguyen.
+ * Email: khang.nguyen@banvien.com
+ * Date: 09/03/2019
+ * Time: 11:26 AM
+ */
+@ExtendWith(SpringExtension.class)
+@SpringBootTest
+public class PasswordEncodeTest {
+    @Test
+    public void testPassword() {
+        BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
+        String pw = encoder.encode("123456");
+        System.out.println("[ "+ pw + " ]");
+    }
+}
