+++ src/test/java/com/poseidon/tradingapp/RuleTests.java
@@ -1,0 +1,47 @@
+package com.poseidon.tradingapp;
+
+import com.poseidon.tradingapp.domain.RuleName;
+import com.poseidon.tradingapp.repositories.RuleNameRepository;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+import java.util.List;
+import java.util.Optional;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@ExtendWith(SpringExtension.class)
+@SpringBootTest
+public class RuleTests {
+
+	@Autowired
+	private RuleNameRepository ruleNameRepository;
+
+	@Test
+	public void ruleTest() {
+		RuleName rule = new RuleName("Rule Name", "Description", "Json", "Template", "SQL", "SQL Part");
+
+		// Save
+		rule = ruleNameRepository.save(rule);
+		assertNotNull(rule.getId());
+        assertEquals("Rule Name", rule.getName());
+
+		// Update
+		rule.setName("Rule Name Update");
+		rule = ruleNameRepository.save(rule);
+        assertEquals("Rule Name Update", rule.getName());
+
+		// Find
+		List<RuleName> listResult = ruleNameRepository.findAll();
+        assertFalse(listResult.isEmpty());
+
+		// Delete
+		Integer id = rule.getId();
+		ruleNameRepository.delete(rule);
+		Optional<RuleName> ruleList = ruleNameRepository.findById(id);
+		assertFalse(ruleList.isPresent());
+	}
+}
